chore: move from io.github.metacosm to net.laprun.sustainability
