+++ metadata/pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>io.github.metacosm</groupId>
+        <groupId>net.laprun.sustainability</groupId>
         <artifactId>power-server-parent</artifactId>
         <version>1.0.0-SNAPSHOT</version>
     </parent>
@@ -20,7 +20,7 @@
                     <artifactId>formatter-maven-plugin</artifactId>
                     <dependencies>
                         <dependency>
-                            <groupId>io.github.metacosm</groupId>
+                            <groupId>net.laprun.sustainability</groupId>
                             <artifactId>build-tools</artifactId>
                             <version>1.0.0-SNAPSHOT</version>
                         </dependency>
