--- metadata/src/main/java/io/github/metacosm/power/SensorMeasure.java
@@ -1,4 +1,0 @@
-package io.github.metacosm.power;
-
-public record SensorMeasure(double[] components, long tick) {
-}
