+++ metadata/src/main/java/net/laprun/sustainability/power/SensorMeasure.java
@@ -1,0 +1,4 @@
+package net.laprun.sustainability.power;
+
+public record SensorMeasure(double[] components, long tick) {
+}
