--- server/src/main/java/io/github/metacosm/power/sensors/Measures.java
@@ -1,21 +1,0 @@
-package io.github.metacosm.power.sensors;
-
-import java.util.Set;
-
-import io.github.metacosm.power.SensorMeasure;
-
-public interface Measures {
-    SensorMeasure missing = new SensorMeasure(new double[] { -1.0 }, -1);
-
-    RegisteredPID register(long pid);
-
-    void unregister(RegisteredPID registeredPID);
-
-    Set<RegisteredPID> trackedPIDs();
-
-    int numberOfTrackerPIDs();
-
-    void record(RegisteredPID pid, SensorMeasure sensorMeasure);
-
-    SensorMeasure getOrDefault(RegisteredPID pid);
-}
