--- server/src/main/java/io/github/metacosm/power/sensors/RegisteredPID.java
@@ -1,12 +1,0 @@
-package io.github.metacosm.power.sensors;
-
-public record RegisteredPID(String stringForMatching) {
-
-    public RegisteredPID(long pid) {
-        this(prepare(pid));
-    }
-
-    public static String prepare(long pid) {
-        return " " + pid + " ";
-    }
-}
