--- server/src/main/java/io/github/metacosm/power/sensors/linux/rapl/RAPLFile.java
@@ -1,11 +1,0 @@
-package io.github.metacosm.power.sensors.linux.rapl;
-
-import java.nio.file.Path;
-
-interface RAPLFile {
-    long extractPowerMeasure();
-
-    static RAPLFile createFrom(Path file) {
-        return ByteBufferRAPLFile.createFrom(file);
-    }
-}
