+++ server/src/main/java/net/laprun/sustainability/power/sensors/RegisteredPID.java
@@ -1,0 +1,12 @@
+package net.laprun.sustainability.power.sensors;
+
+public record RegisteredPID(String stringForMatching) {
+
+    public RegisteredPID(long pid) {
+        this(prepare(pid));
+    }
+
+    public static String prepare(long pid) {
+        return " " + pid + " ";
+    }
+}
