--- server/src/test/java/io/github/metacosm/power/CIQuarkusTestProfile.java
@@ -1,13 +1,0 @@
-package io.github.metacosm.power;
-
-import java.util.Set;
-
-import io.quarkus.test.junit.QuarkusTestProfile;
-
-public class CIQuarkusTestProfile implements QuarkusTestProfile {
-
-    @Override
-    public Set<String> tags() {
-        return Set.of("ci");
-    }
-}
