--- server/src/test/java/io/github/metacosm/power/PowerResourceIT.java
@@ -1,8 +1,0 @@
-package io.github.metacosm.power;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-
-@QuarkusIntegrationTest
-public class PowerResourceIT extends PowerResourceTest {
-    // Execute the same tests but in packaged mode.
-}
