+++ server/src/test/java/net/laprun/sustainability/power/MockPowerMeasurer.java
@@ -1,0 +1,13 @@
+package net.laprun.sustainability.power;
+
+import io.quarkus.test.Mock;
+
+@Mock
+@SuppressWarnings("unused")
+public class MockPowerMeasurer extends PowerMeasurer {
+
+    @Override
+    protected long validPIDOrFail(String pid) {
+        return Long.parseLong(pid);
+    }
+}
