+++ server/src/test/java/net/laprun/sustainability/power/PowerResourceIT.java
@@ -1,0 +1,8 @@
+package net.laprun.sustainability.power;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class PowerResourceIT extends PowerResourceTest {
+    // Execute the same tests but in packaged mode.
+}
